Merge `stdout` and `stderr` into just `stdout`:
Turns out that `stdout` and `stderr` weren't being captured for the container
executor, and values weren't being printed to the screen with the local
executor. In fixing this I realized that storing the stdout/stderr separately
would be too much of a hassle, and would make it impossible to recombine the
stdout/stderr back together. I might change this approach later, but for now
this will do.