Scope terminal sessions to workflows instead of sessions:
Sessions and terminal sessions have had a long and weird history, and managing
the IDs has been a pain. Previously a session could not be reran, so a single
terminal session could be assigned to a session no problem. Then, when I added
the ability to rerun sessions, I needed a way to create a unique id for the
terminal since the session id was still the same. The hack I made was
concatenating the session and run number, which was not ideal. Now I am
switching to using workflows as the 1-1 mapping between terminal sessions and
workflows. This should make managing terminals/sessions/workflows easier going
forwards.