Add list of workflows to sessions:
I've decided the simplest way to represent sessions is to make them a list of
workflows. When a session is reran, it will add a new workflow to the session.
In this system there will only be one session object, and the run number will
be determined from the index of the workflow in the session object.

Secondly, I've also figured out how to "easily" add matrix support: Each
workflow will be able to specify a parent workflow, meaning that any single
workflow will now potentially have a list of sub-workflows that it spawned.
Wiring this up shouldn't be too difficult, and makes it easy down the line to
add wire diagrams or graph displays of pipelines.