Add button to show workflow file that ran workflow:
This is a bit hacky, but should be made more elegant when I am finished with
all the workflow/session migration stuff. In short, from now on workflows will
have a reference to the filename that is being used to run the workflow. To get
this to work I had to do a bunch of hacky stuff on the backend, mainly fixing
some of the shortcuts I made for the integration tests. Also, the "get session
by session id" endpoint now returns a session with all of the workflows for
that session populated. This shouldn't be a huge concern, though it is
ineffecient. Since workflows have the same permissions as sessions, there
shouldn't be any issue with this, though it is something to keep in mind for
future refactors. Ideally the frontend would only need to request the workflow
that it needs, but since it is still using the "session" methodology this will
have to do.