Add basic function annotation support:
This is a stepping stone to allow for "matrix builds", or in Cicada language,
sub-workflows. Essentially, users will be able to annotate a function with the
`@workflow` annotation, and when that function is called, a new workflow will
be spawned/forked from the existing one and ran in parallel.

Currently annotations are not first class citizens like they are in Python or
Kotlin, meaning you cannot create your own annotations just yet. The only
annotation you can use is `@workflow`, and even that annotation has yet to be
implemented. Better annotation support is coming in the future though!