Disable iframe embedding:
Previously the website was able to be clickjacked, meaning an attacker could
embed a page in an iframe via a website/email and phish a user into clicking
or typing information into the website while secretly intercepting it. This
has been fixed by disabling the site from being embeded in an iframe.