Add Gitlab SSO support:
This allows users to signup/login with their Gitlab accounts. This doesn't
allow for Gitlab users to connect webhooks to Cicada, that will come in a
future update.

To make this work I needed to create a custom Gitlab OAuth application which
allows users to authenticate with it. The permissions have yet to be ironed
out, and currently it is probably asking for too many permissions (all related
to user profile/email info). This is needed to get the user's email and
username so that a user profile can be created. I plan on experimenting with
OpenID Connect as an alternative to using the Gitlab API for getting user data
which is why I'm leaving the permissions as is for now.

I also made some backend/UI changes to get this working:

* Add a Gitlab login button to the login screen

* Change the login screen to use white buttons with logos for 3rd party
  identity providers (Github/Gitlab)

* Replace `get_user_by_username` with `get_user_by_username_and_provider`.
  Previously it was assumed that you where either using `cicada` or `github` as
  a provider, so there was lots of places where this logic was either assumed
  or hard coded. Now you have to explictly pass the provider to be able to get
  a user by their username. This should prevent any name collisions where a
  user with the same username for a different provider is returned.