Store Gitlab OAuth tokens in DB:
Because Gitlab's API doesn't support authenticating via OAuth applications like
GitHub does, we have to store OAuth the access/refresh tokens when a user logs
in via SSO. The API keys are needed in order to make API calls on behalf of the
user, which is needed in order to setup webhooks automatically. The alternative
would be to make the user re-authenticate via Gitlab each time they want to do
an API request to Gitlab (annoying for the user), or store the token on the
client browser and make requests there/pass it to the server (insecure, since
the token has `api` scope, which is needed for making certain API calls).

The best route is to store the tokens server side. The current flow is like
this: When a login occurs the new token for the user is stored in-memory
unencrypted, and then stored in the database encrypted using Fernet symmetric
cryptography from the `cryptography` Python package. Then, when the server
wants a token for a given user, the in-memory cache is checked, and if it is
not found, it is pulled from the DB. In either case, if the token has expired,
the refresh token is used to grab a new access/refresh token, then these new
values are stored like before, and returned to the caller.

This is probably overkill, but should handle (most) cases when it comes to
using the Gitlab API with OAuth authenticated users.