Make AST visitor async:
Normally walking an AST tree is not CPU/IO intensive, but because certain nodes
make IO calls (like `cache` and `shell`), some sort of async calling is needed.
The old approach was to execute the AST tree in a separate thread/process,
meaning a whole thread/process would be created for each workflow run, and each
of those workflows would spawn subprocesses for each `shell` call, which is not
very effecient. And, in order to better support sub-workflows, a more managable
system is needed for creating ad-hoc workflows while in the middle of walking
an existing AST tree. Having everything (or most things) using this new async
model means less IPC and less threads/processes down the line.