Add UI for sub workflows:
This adds the frontend code for the new sub workflow support I added to the
backend. More specifically, when a sub workflow workflow is created, a new line
appears with the title, status, and elapsed time. Clicking on the line will
replace the current terminal (for the root workflow) with the terminal output
of the sub workflow. This allows you to toggle/switch between the different
terminal outputs without navigating away (or viewing all the output on one
screen like before). To facilitate these changes I had to change the backend
as well to send the workflow ids associated with the stdout/status changes
that are occurring, which means I now need to poll the DB for workflow changes.
This shouldn't be an issue, though it might need to be optimized in the future.